docs(styles): split button stories [ci visual]
