+++ packages/styles/stories/Components/button-split/split-menu-button.example.html
@@ -61,4 +61,23 @@
             <i class="sap-icon--slim-arrow-down"></i>
         </button>
     </div>
+
+    <div class="fd-button-split fd-button-split--attention" role="group" aria-label="button-split">
+        <button class="fd-button fd-button--attention" aria-label="button">
+            <i class="sap-icon--cart"></i>
+            <span class="fd-button-split__text">Attention</span>
+        </button>
+        <button class="fd-button fd-button--attention" aria-expanded="false">
+            <i class="sap-icon--slim-arrow-down"></i>
+        </button>
+    </div>
+
+    <div class="fd-button-split fd-button-split--negative" role="group" aria-label="button-split">
+        <button class="fd-button fd-button--negative" aria-label="button">
+            <i class="sap-icon--cart"></i>
+        </button>
+        <button class="fd-button fd-button--negative" aria-expanded="false">
+            <i class="sap-icon--slim-arrow-down"></i>
+        </button>
+    </div>
 </div>
