+++ packages/styles/stories/Components/button/icon-and-text.example.html
@@ -1,18 +1,75 @@
 
     <div class="fddocs-button-container">
-        <button class="fd-button fd-button--emphasized">Add to Cart</button>
-        <button class="fd-button">Add to Cart</button>
-        <button class="fd-button fd-button--transparent">Add to Cart</button>
-        <button class="fd-button fd-button--ghost">Add to Cart</button>
-        <button class="fd-button fd-button--positive">Approve</button>
-        <button class="fd-button fd-button--negative">Reject</button>
-        <button class="fd-button fd-button--attention">Attention</button>
+        <button class="fd-button fd-button--emphasized">
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button class="fd-button">
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button class="fd-button fd-button--transparent">
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button class="fd-button fd-button--ghost">
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button class="fd-button fd-button--positive">
+            <span class="fd-button__text">Approve</span>
+        </button>
+        <button class="fd-button fd-button--negative">
+            <span class="fd-button__text">Reject</span>
+        </button>
+        <button class="fd-button fd-button--attention">
+            <span class="fd-button__text">Attention</span>
+        </button>
         <br><br>
-        <button aria-label="Add to cart" class="fd-button fd-button--emphasized"><i class="sap-icon--cart"></i></button>
-        <button aria-label="Add to cart" class="fd-button"><i class="sap-icon--cart"></i></button>
-        <button aria-label="Add to cart" class="fd-button fd-button--transparent"><i class="sap-icon--cart"></i></button>
-        <button aria-label="Filter" class="fd-button fd-button--ghost"><i class="sap-icon--filter"></i></button>
-        <button aria-label="Accept" class="fd-button fd-button--positive"><i class="sap-icon--accept"></i></button>
-        <button aria-label="Decline" class="fd-button fd-button--negative"><i class="sap-icon--decline"></i></button>
-        <button aria-label="Decline" class="fd-button fd-button--attention"><i class="sap-icon--decline"></i></button>
+        <button aria-label="Add to cart" class="fd-button fd-button--emphasized">
+            <i class="sap-icon--cart"></i>
+        </button>
+        <button aria-label="Add to cart" class="fd-button">
+            <i class="sap-icon--cart"></i>
+        </button>
+        <button aria-label="Add to cart" class="fd-button fd-button--transparent">
+            <i class="sap-icon--cart"></i>
+        </button>
+        <button aria-label="Filter" class="fd-button fd-button--ghost">
+            <i class="sap-icon--filter"></i>
+        </button>
+        <button aria-label="Accept" class="fd-button fd-button--positive">
+            <i class="sap-icon--accept"></i>
+        </button>
+        <button aria-label="Decline" class="fd-button fd-button--negative">
+            <i class="sap-icon--decline"></i>
+        </button>
+        <button aria-label="Decline" class="fd-button fd-button--attention">
+            <i class="sap-icon--decline"></i>
+        </button>
+        <br><br>
+        <button aria-label="Add to cart" class="fd-button fd-button--emphasized">
+            <i class="sap-icon--cart"></i>
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button aria-label="Add to cart" class="fd-button">
+            <i class="sap-icon--cart"></i>
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button aria-label="Add to cart" class="fd-button fd-button--transparent">
+            <i class="sap-icon--cart"></i>
+            <span class="fd-button__text">Add to Cart</span>
+        </button>
+        <button aria-label="Filter" class="fd-button fd-button--ghost">
+            <i class="sap-icon--filter"></i>
+            <span class="fd-button__text">Filter</span>
+        </button>
+        <button aria-label="Accept" class="fd-button fd-button--positive">
+            <i class="sap-icon--accept"></i>
+            <span class="fd-button__text">Accept</span>
+        </button>
+        <button aria-label="Decline" class="fd-button fd-button--negative">
+            <i class="sap-icon--decline"></i>
+            <span class="fd-button__text">Decline</span>
+        </button>
+        <button aria-label="Decline" class="fd-button fd-button--attention">
+            <i class="sap-icon--decline"></i>
+            <span class="fd-button__text">Decline</span>
+        </button>
     </div>
