+++ packages/styles/stories/Components/button/styles.example.html
@@ -1,9 +1,23 @@
 <div class="fddocs-container fddocs-button-container">
-    <button class="fd-button">Default Button</button>
-    <button class="fd-button fd-button--emphasized">Emphasized Button</button>
-    <button class="fd-button fd-button--ghost">Ghost Button</button>
-    <button class="fd-button fd-button--positive">Positive Button</button>
-    <button class="fd-button fd-button--negative">Negative Button</button>
-    <button class="fd-button fd-button--attention">Attention Button</button>
-    <button class="fd-button fd-button--transparent">Transparent Button</button>
+    <button class="fd-button">
+        <span class="fd-button__text">Default Button</span>
+    </button>
+    <button class="fd-button fd-button--emphasized">
+        <span class="fd-button__text">Emphasized Button</span>
+    </button>
+    <button class="fd-button fd-button--ghost">
+        <span class="fd-button__text">Ghost Button</span>
+    </button>
+    <button class="fd-button fd-button--positive">
+        <span class="fd-button__text">Positive Button</span>
+    </button>
+    <button class="fd-button fd-button--negative">
+        <span class="fd-button__text">Negative Button</span>
+    </button>
+    <button class="fd-button fd-button--attention">
+        <span class="fd-button__text">Attention Button</span>
+    </button>
+    <button class="fd-button fd-button--transparent">
+        <span class="fd-button__text">Transparent Button</span>
+    </button>
 </div>
